pyproject.toml
[tool.poetry]
name = "quibe-bot"
version = "0.1.0"
description = ""
authors = ["caiopo <caiopoliveira@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "=10.1.0"
dataset = "=1.0.6"
libru = "^0.1.2"
normality = "=0.6.1"
pytz = "^2018.9"
psycopg2-binary = "^2.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
